register new game API endpoint implemented
